+++ index.html
@@ -14,12 +14,14 @@
       <ul id="type-list">
       </ul>
 
-      <label class="switch">
-        <input type="checkbox">
-        <span class="slider round">
-        </span>
+      <div class="switcher-container">
+        <label class="switch">
+          <input id="toggleCheckbox" type="checkbox">
+          <span class="slider round">
+          </span>
+        </label>
         <span>Vege only</span>
-      </label>
+      </div>
 
     </div>
     <section class="container-categories">
@@ -31,6 +33,7 @@
   <div id="app"></div>
   <script type="module" src="/main.js"></script>
   <script type="module" src="/scripts/contactDb.js"></script>
+  <script type="module" src="/scripts/generalScript.js"></script>
 </body>
 
 </html>
