+++ scripts/shared/const.js
@@ -1,0 +1,29 @@
+
+export const productsUrl = "https://api-eko-bazarek.azurewebsites.net/api/products/categories";
+export const typesUrl = "https://api-eko-bazarek.azurewebsites.net/api/products/types";
+
+export const vegeProducts = [{
+    id: 'HONEY',
+    name: 'Miody'
+},
+{
+    id: 'DAIRY',
+    name: 'Nabiał'
+},
+{
+    id: 'FRUITS',
+    name: 'Owoce'
+},
+{
+    id: 'VEGETABLE',
+    name: 'Warzywa'
+},
+{
+    id: 'CEREALS',
+    name: 'Zboża'
+},
+{
+    id: 'WINE_AND_SPIRITS',
+    name: 'Alkohole'
+}
+];
