+++ style.css
@@ -55,6 +55,15 @@ ul {
   color: #eeeeee;
 }
 
+.switcher-container {
+  display: flex;
+  align-items: center;
+}
+
+div > span {
+  margin-left: 1rem;
+}
+
 .container-categories {
   margin-left: 1rem;
 }
